+++ src/pyo3_extensions.rs
@@ -4,10 +4,14 @@ use crate::try_unwrap;
 
 use pyo3::basic::CompareOp;
 use pyo3::exceptions::PyTypeError;
+use pyo3::intern;
 use pyo3::prelude::*;
-use pyo3::types::*;
+use pyo3::sync::GILOnceCell;
+use pyo3::types::PyBytes;
 use serde::ser::Error;
 use std::fmt;
+
+static PICKLE_MODULE: GILOnceCell<Py<PyModule>> = GILOnceCell::new();
 
 /// Represents a Python object flowing through a Timely dataflow.
 ///
@@ -60,7 +64,7 @@ impl std::fmt::Debug for TdPyAny {
 }
 
 /// Serialize Python objects flowing through Timely that cross
-/// process bounds as pickled bytes.
+/// process bounds as bytes.
 impl serde::Serialize for TdPyAny {
     // We can't do better than isolating the Result<_, PyErr> part and
     // the explicitly converting.  1. `?` automatically trys to
@@ -84,9 +88,14 @@ impl serde::Serialize for TdPyAny {
     {
         Python::with_gil(|py| {
             let x = self.bind(py);
-            let pickle = py.import_bound("pickle").map_err(S::Error::custom)?;
+            let pickle = PICKLE_MODULE
+                .get_or_try_init(py, || -> PyResult<Py<PyModule>> {
+                    Ok(py.import_bound("pickle")?.unbind())
+                })
+                .map_err(S::Error::custom)?;
             let binding = pickle
-                .call_method1("dumps", (x,))
+                .bind(py)
+                .call_method1(intern!(py, "dumps"), (x,))
                 .map_err(S::Error::custom)?;
             let bytes = binding.downcast::<PyBytes>().map_err(S::Error::custom)?;
             serializer
@@ -111,7 +120,10 @@ impl<'de> serde::de::Visitor<'de> for PickleVisitor {
     {
         let x: Result<TdPyAny, PyErr> = Python::with_gil(|py| {
             let pickle = py.import_bound("pickle")?;
-            let x = pickle.call_method1("loads", (bytes,))?.unbind().into();
+            let x = pickle
+                .call_method1(intern!(py, "loads"), (bytes,))?
+                .unbind()
+                .into();
             Ok(x)
         });
         x.map_err(E::custom)
@@ -124,45 +136,6 @@ impl<'de> serde::Deserialize<'de> for TdPyAny {
     fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
         deserializer.deserialize_bytes(PickleVisitor)
     }
-}
-
-// Rust tests that interact with the Python interpreter don't work
-// well under pyenv-virtualenv. This test executes under the global pyenv
-// version, instead of the configured virtual environment.
-// Disabling this test for aarch64, as it fails in CI.
-#[cfg(not(target_arch = "aarch64"))]
-#[test]
-fn test_serde() {
-    use serde_test::assert_tokens;
-    use serde_test::Token;
-
-    pyo3::prepare_freethreaded_python();
-
-    let pyobj: TdPyAny =
-        Python::with_gil(|py| PyString::new_bound(py, "hello").into_any().unbind().into());
-
-    // Python < 3.8 serializes strings differently than python >= 3.8.
-    // We get the current python version here so we can assert based on that.
-    let (major, minor) = Python::with_gil(|py| {
-        let sys = PyModule::import_bound(py, "sys").unwrap();
-        let version = sys.getattr("version_info").unwrap();
-        let major: i32 = version.getattr("major").unwrap().extract().unwrap();
-        let minor: i32 = version.getattr("minor").unwrap().extract().unwrap();
-        (major, minor)
-    });
-
-    // We only support python 3...
-    assert_eq!(major, 3);
-
-    let expected = if minor < 8 {
-        Token::Bytes(&[128, 3, 88, 5, 0, 0, 0, 104, 101, 108, 108, 111, 113, 0, 46])
-    } else {
-        Token::Bytes(&[
-            128, 4, 149, 9, 0, 0, 0, 0, 0, 0, 0, 140, 5, 104, 101, 108, 108, 111, 148, 46,
-        ])
-    };
-    // This does a round-trip.
-    assert_tokens(&pyobj, &[expected]);
 }
 
 /// Re-use Python's value semantics in Rust code.